moved thug to specific integration
