+++ integrations/malware_tools_analyzers/Dockerfile
@@ -1,6 +1,7 @@
-# this base was derived from Thug requirements:
-# https://github.com/REMnux/docker/blob/master/thug/Dockerfile
-FROM python:3.9-slim
+FROM python:3.11-slim
+
+ARG TARGETARCH
+RUN echo "TargetArch is $TARGETARCH"
 
 ENV PROJECT_PATH=/opt/deploy
 ENV LOG_PATH=/var/log/intel_owl/malware_tools_analyzers
@@ -9,17 +10,14 @@ ENV USER=malware_tools_analyzers-user
 # update and install packages
 # line 3: ClamAV deps
 # line 4: Box-JS deps
-# lines 5-6: Thug deps
 RUN DEBIAN_FRONTEND=noninteractive apt-get update -qq \
     && apt-get install -y --no-install-recommends wget git libssl3 swig g++ make libssl-dev libmagic1 vim \
     clamav clamdscan clamav-daemon clamav-freshclam \
-    nodejs npm gcc m4 \
-    build-essential python3-dev python3-setuptools python3-wheel python-is-python3 libboost-dev libboost-iostreams-dev libboost-python-dev libboost-system-dev python3-pip libxml2-dev \
-    libxslt-dev tesseract-ocr unzip libtool graphviz-dev automake libffi-dev graphviz libfuzzy-dev libfuzzy2 libjpeg-dev libffi-dev pkg-config clang autoconf
+    nodejs npm gcc m4
 
 # Add a new low-privileged user
 RUN useradd -ms /bin/bash ${USER} \
-    && mkdir ${PROJECT_PATH} ${PROJECT_PATH}/thug ${PROJECT_PATH}/qiling ${PROJECT_PATH}/stringsifter ${PROJECT_PATH}/peframe ${PROJECT_PATH}/apkid
+    && mkdir ${PROJECT_PATH} ${PROJECT_PATH}/qiling ${PROJECT_PATH}/stringsifter ${PROJECT_PATH}/peframe ${PROJECT_PATH}/apkid
 
 WORKDIR ${PROJECT_PATH}
 
@@ -61,10 +59,11 @@ RUN python3 -m venv venv \
 # Build Qiling
 WORKDIR ${PROJECT_PATH}/qiling
 COPY requirements/qiling-requirements.txt qiling/analyze.py ./
-RUN python3 -m venv venv \
+# keystone-engine does not compile for ARM
+RUN if [[ $TARGETARCH == "amd" ]]; then python3 -m venv venv \
     && . venv/bin/activate \
     && pip3 install --no-cache-dir --upgrade pip \
-    && pip3 install --no-cache-dir -r qiling-requirements.txt
+    && pip3 install --no-cache-dir -r qiling-requirements.txt fi
 
 # Build APKiD
 WORKDIR ${PROJECT_PATH}/apkid
@@ -130,27 +129,6 @@ RUN python3 -m venv venv \
     && mkdir -p /root/.semgrep/ \
     && chown -R ${USER}:${USER} /root/.semgrep \
     && chmod 711 /root
-
-# Install Thug
-# https://github.com/buffer/thug/blob/master/docker/Dockerfile
-WORKDIR ${PROJECT_PATH}/thug
-COPY requirements/thug-requirements.txt ./
-RUN python3 -m venv venv \
-    && . venv/bin/activate \
-    && pip3 install --no-cache-dir --upgrade pip \
-    # this is the python 3.9 version. Once you update the base image you should switch to the right version (cp39)
-    && wget -q https://github.com/cloudflare/stpyv8/releases/download/v12.7.224.18/stpyv8-12.7.224.18-cp39-cp39-manylinux_2_31_x86_64.whl \
-    && pip3 install --no-cache-dir stpyv8-12.7.224.18-cp39-cp39-manylinux_2_31_x86_64.whl \
-    && rm stpyv8-12.7.224.18-cp39-cp39-manylinux_2_31_x86_64.whl \
-    && mkdir -p /usr/share/stpyv8 \
-    && git clone https://github.com/buffer/libemu.git && cd libemu && autoreconf -v -i && ./configure && make install && cd .. && rm -rf libemu && ldconfig \
-    && pip3 install --no-cache-dir -r thug-requirements.txt \
-    && git clone --depth 1 https://github.com/buffer/thug.git \
-    && mkdir -p /etc/thug \
-    && cp -R thug/thug/conf/* /etc/thug \
-    && rm -rf thug \
-    && mkdir -p /tmp/thug/logs \
-    && chown -R ${USER}:${USER} /tmp/thug/logs /etc/thug
 
 # prepare fangfrisch installation
 COPY crontab /etc/cron.d/crontab
