+++ integrations/malware_tools_analyzers/app.py
@@ -130,44 +130,6 @@ def intercept_droidlysis_result(context, future: Future) -> None:
         shutil.rmtree(dir_loc, ignore_errors=True)
 
 
-def intercept_thug_result(context, future: Future) -> None:
-    """
-    Thug doesn't output result to standard output but to a file,
-    using this callback function,
-    we intercept the future object and update its result attribute
-    by reading the final analysis result from the saved result file
-    before it is ready to be consumed.
-    """
-    # 1. get current result object
-    res = future.result()
-    # 2. dir from which we will read final analysis result
-    dir_loc = context.get("read_result_from", None)
-    if not dir_loc:
-        res["error"] += ", No specified file to read result from"
-        if res.get("returncode", -1) == 0:
-            res["returncode"] = -1
-    else:
-        # 3. read saved result file, if it exists
-        f_loc = dir_loc + "/analysis/json/analysis.json"
-        if not os.path.exists(f_loc):
-            res["error"] += f", result file {f_loc} does not exists."
-            if res.get("returncode", -1) == 0:
-                res["returncode"] = -1
-        else:
-            with open(f_loc, "r", encoding="utf-8") as fp:
-                try:
-                    res["report"] = json.load(fp)
-                except json.JSONDecodeError:
-                    res["report"] = fp.read()
-
-    # 4. set final result after modifications
-    future._result = res  # skipcq PYL-W0212
-
-    # 5. directory can be removed now
-    if dir_loc:
-        shutil.rmtree(dir_loc, ignore_errors=True)
-
-
 # with this, we can make http calls to the endpoint: /capa
 shell2http.register_command(endpoint="capa", command_name="/usr/local/bin/capa -q -j")
 
@@ -238,11 +200,4 @@ shell2http.register_command(
 shell2http.register_command(
     endpoint="goresym",
     command_name="/usr/local/bin/goresym",
-)
-
-# with this, we can make http calls to the endpoint: /thug
-shell2http.register_command(
-    endpoint="thug",
-    command_name="/opt/deploy/thug/venv/bin/thug -qZF",
-    callback_fn=intercept_thug_result,
 )
