+++ integrations/malware_tools_analyzers/requirements/qiling-requirements.txt
@@ -1,1 +1,1 @@
-qiling==1.4.5
+qiling==1.4.6
