--- integrations/malware_tools_analyzers/requirements/thug-requirements.txt
@@ -1,7 +1,0 @@
-setuptools==70.0.0
-pytesseract==0.3.10
-# this has to be pinned to 1.10 otherwise the image build will fail
-pygraphviz==1.10
-# CAREFUL! This is strictly tied to STpy version and Python version
-# DO NOT UPGRADE THIS WITHOUT PROPER MANUAL TESTING
-thug==4.9
