--- requirements/pre-requirements.txt
@@ -1,2 +1,0 @@
-python-dotenv>=0.21.0
-GitPython>=3.1.30
